Move the TensorFlow tests under a `tensorflow.v2` package and rename the base to `AbstractTensorTest` (wala/ML#635).
The framework and API-generation identity belongs in the package, not the class
names: `com.ibm.wala.cast.python.ml.test.tensorflow.v2` carries both, so the base
no longer needs `TensorFlow` in its name (it was redundant with the package) and
neither will the feature-area classes. `v1` is reserved for the existing TF1
classes, and `tensorflow` sits beside future `jax`/`pytorch` packages. The base
`AbstractTensorFlowTest` becomes `AbstractTensorTest`. Behavior preserving: the
suite passes with the same test count.
