+++ com.ibm.wala.cast.python.ml.test/source/com/ibm/wala/cast/python/ml/test/DiagnosticLoggingVolumeTest.java
@@ -2,7 +2,7 @@ package com.ibm.wala.cast.python.ml.test;
 
 import static org.junit.Assert.assertTrue;
 
-import com.ibm.wala.cast.python.ml.test.tensorflow.TestTensorflow2Model;
+import com.ibm.wala.cast.python.ml.test.tensorflow.v2.TestTensorflow2Model;
 import com.ibm.wala.ipa.cha.ClassHierarchyException;
 import com.ibm.wala.util.CancelException;
 import java.io.IOException;
